moved insert utils to its own module
